# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Base Plate API",
        "description": "A product of Appetiser",
        "contact": {
            "name": "API Support",
            "email": "dennis.agulo@appetiser.com.au"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://128.199.193.26/api/v1",
            "description": "Base Endpoint"
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login user. Authenticate credentials and returns a token.",
                "operationId": "App\\Http\\Controllers\\API\\V1\\ApiAuthController::login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Email Address",
                                        "type": "string",
                                        "example": "testuser@gmail.com"
                                    },
                                    "password": {
                                        "description": "Password",
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Invalid Credentials"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "200": {
                        "description": "Login Success"
                    }
                },
                "security": []
            }
        },
        "/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register user. Creates new user, sends 'thank you' email and returns a token.",
                "operationId": "App\\Http\\Controllers\\API\\V1\\ApiAuthController::register",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Email Address",
                                        "type": "string",
                                        "example": "testuser@gmail.com"
                                    },
                                    "password": {
                                        "description": "Password",
                                        "type": "string",
                                        "example": "password"
                                    },
                                    "password_confirmation": {
                                        "description": "Confirm Password",
                                        "type": "string",
                                        "example": "password"
                                    },
                                    "first_name": {
                                        "description": "First Name",
                                        "type": "string",
                                        "example": "Jane"
                                    },
                                    "last_name": {
                                        "description": "Last Name",
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "dob": {
                                        "description": "Date of Birth (Y-m-d)",
                                        "type": "date",
                                        "example": "1991-01-01"
                                    },
                                    "country": {
                                        "description": "Country",
                                        "type": "string",
                                        "example": "Australia"
                                    },
                                    "mobile_number": {
                                        "description": "Mobile Number",
                                        "type": "string",
                                        "example": "+61 412345678"
                                    },
                                    "address": {
                                        "description": "Address",
                                        "type": "string",
                                        "example": "85 Dover Street Melbourne VIC"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Invalid Input"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "201": {
                        "description": "Request Created"
                    }
                },
                "security": []
            }
        },
        "/auth/email/check": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Check email if exists. Used as a first step in user registration",
                "description": "",
                "operationId": "",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/auth/user": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get authenticated user's information",
                "operationId": "App\\Http\\Controllers\\API\\V1\\ApiAuthController::getAuthUser",
                "responses": {
                    "400": {
                        "description": "Invalid Token"
                    },
                    "401": {
                        "description": "Token Expired"
                    },
                    "404": {
                        "description": "Token Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "200": {
                        "description": "Authenticated User"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/auth/logout": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout user",
                "operationId": "App\\Http\\Controllers\\API\\V1\\ApiAuthController::logout",
                "responses": {
                    "400": {
                        "description": "Invalid Token"
                    },
                    "401": {
                        "description": "Token Expired"
                    },
                    "404": {
                        "description": "Token Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "200": {
                        "description": "Logout Success"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}